Suppress -Wsign-compare via config instead of (int) casts
Agent-Logs-Url: https://github.com/blue-jam/ProconLibrary/sessions/8890298c-e75d-4b6e-9d35-8ec16a4b1798

+++ geometry/arrangement.hpp
@@ -19,18 +19,18 @@ bool merge_if_able(Segment& s, Segment t) {
 }
 // オーバーラップをマージする．（破壊的）
 void merge_segments(vector<Segment>& ss) {
-    for (int i = 0; i < (int)ss.size(); ++i)
+    for (int i = 0; i < ss.size(); ++i)
         if (lessX(ss[i][1], ss[i][0])) swap(ss[i][1], ss[i][0]);
-    for (int i = 0; i < (int)ss.size(); ++i)
-        for (int j = i + 1; j < (int)ss.size(); ++j)
+    for (int i = 0; i < ss.size(); ++i)
+        for (int j = i + 1; j < ss.size(); ++j)
             if (merge_if_able(ss[i], ss[j]))
                 ss[j--] = ss.back(), ss.pop_back();
 }
 Graph<double> segmentArrangement(const vector<Segment>& ss, vector<P>& ps) {
-    for (int i = 0; i < (int)ss.size(); ++i) {
+    for (int i = 0; i < ss.size(); ++i) {
         ps.push_back(ss[i][0]);
         ps.push_back(ss[i][1]);
-        for (int j = i + 1; j < (int)ss.size(); ++j) {
+        for (int j = i + 1; j < ss.size(); ++j) {
             if (intersectSS(ss[i], ss[j]))
                 ps.push_back(crosspointSS(ss[i], ss[j]));
         }
@@ -38,14 +38,14 @@ Graph<double> segmentArrangement(const vector<Segment>& ss, vector<P>& ps) {
     sort(ALL(ps), lessX);
     ps.erase(unique(ALL(ps), near), ps.end());
     Graph<double> g(ps.size());
-    for (int i = 0; i < (int)ss.size(); ++i) {
+    for (int i = 0; i < ss.size(); ++i) {
         vector<pair<double, int>> lst;
-        for (int j = 0; j < (int)ps.size(); ++j) {
+        for (int j = 0; j < ps.size(); ++j) {
             if (ccw(ss[i][0], ss[i][1], ps[j]) == 0)
                 lst.push_back(make_pair(norm(ss[i][0] - ps[j]), j));
         }
         sort(ALL(lst));
-        for (int j = 0; j + 1 < (int)lst.size(); ++j) {
+        for (int j = 0; j + 1 < lst.size(); ++j) {
             int a = lst[j].second, b = lst[j + 1].second;
             addUndirectedEdge(g, a, b, abs(ps[a] - ps[b]));
         }
@@ -53,14 +53,14 @@ Graph<double> segmentArrangement(const vector<Segment>& ss, vector<P>& ps) {
     return g;
 }
 bool walkCW(const Graph<double>& g, int v, int u, vector<int>& lst, const vector<P>& ps, vector<vector<int>>& used, vector<vector<int>>& res) {
-    for (int j = 0; j + 1 < (int)lst.size(); ++j)
+    for (int j = 0; j + 1 < lst.size(); ++j)
         if (lst[j] == u && lst[j + 1] == v) {
             Polygon pol;
-            for (int i = j; i < (int)lst.size() - 1; ++i) pol.push_back(ps[lst[i]]);
+            for (int i = j; i < lst.size() - 1; ++i) pol.push_back(ps[lst[i]]);
             double s = area(pol);
             if (s > eps) {
                 vector<int> v;
-                for (int i = j; i < (int)lst.size() - 1; ++i) {
+                for (int i = j; i < lst.size() - 1; ++i) {
                     used[lst[i]][lst[(i + 1) % lst.size()]] = true;
                     v.push_back(lst[i]);
                 }
